pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "level 2 student project"
authors = ["Boris Ageyev <mr.broomy@gmail.com>"]
repository = "https://github.com/mr-xeroth/python-project-lvl2"
packages = [
	{ include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
flake8 = "^4.0"
pytest = "^7.0"
pyyaml = "^5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
